fix the build by moving the test file
The test file was in routes/ so sveltekit bundled it as a route. Then,
when trying to run the app, it realized that `describe` is not defined
and crashed. The solution is either to make this a private module by
prefixing its name with an underscore (https://stackoverflow.com/a/69457826)
or to put them somewhere else - that's what I did.
These are going to be more integration than unit tests anyway, so I
don't mind if they are not next to their prod files.